requirements.txt
absl-py==2.3.1
cffi==2.0.0
clipspy==1.0.6
immutabledict==4.2.2
numpy==2.3.5
ortools==9.14.6206
pandas==2.3.3
protobuf==6.31.1
pycparser==2.23
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.2
